+++ al-khaser/Al-khaser.cpp
@@ -84,6 +84,7 @@ int main(void)
 		exec_check(&cpuid_hypervisor_vendor, TEXT("Checking hypervisor vendor using cpuid(0x40000000)"));
 		exec_check(&accelerated_sleep, TEXT("Check if time has been accelerated: "));
 		exec_check(&VMDriverServices, TEXT("VM Driver Services : "));
+		exec_check(&serial_number_bios_wmi, TEXT("Checking SerialNumber from BIOS: "));
 	}
 
 	/* VirtualBox Detection */