+++ al-khaser/Anti VM/Generic.h
@@ -29,3 +29,5 @@ BOOL accelerated_sleep();
 BOOL cpuid_is_hypervisor();
 BOOL cpuid_hypervisor_vendor();
 BOOL serial_number_bios_wmi();
+BOOL model_computer_system_wmi();
+BOOL manufacturer_computer_system_wmi();